Add recursive dependent questions
