+++ src/Fields/Select/index.js
@@ -1,31 +1,14 @@
 /** @jsx jsx */
-/** @jsxRuntime classic */
-import { jsx } from 'theme-ui'
-import { RHFInput } from 'react-hook-form-input'
-import ReactSelect from 'react-select'
+import { Select as SelectUI, jsx } from 'theme-ui'
 
-const Select = ({
-  register,
-  setValue,
-  name,
-  defaultValue,
-  registerConfig,
-  ...props
-}) => (
-  <RHFInput
-    as={<ReactSelect {...props} />}
-    rules={{
-      ...registerConfig,
-      validate: {
-        noEmpty: (item) =>
-          registerConfig && registerConfig.required ? item.value !== '*' : true
-      }
-    }}
-    name={name}
-    defaultValue={defaultValue}
-    register={register}
-    setValue={setValue}
-  />
-)
+import React from 'react'
+
+const Select = React.forwardRef(({ children, ...props }, ref) => {
+  return (
+    <SelectUI key={ref} ref={ref} {...props}>
+      {children}
+    </SelectUI>
+  )
+})
 
 export default Select
