+++ src/Questions/Select/index.js
@@ -37,53 +37,6 @@ const QuestionSelect = ({
   setValue,
   ...props
 }) => {
-  const DependentQuestion = (dependentQuestion, i) => {
-    return (
-      <div
-        sx={{
-          ...(question.isFullWidth && styles.fullWidth)
-        }}
-        key={i}
-      >
-        {dependentQuestion.condition === watch(question.name) && (
-          <Select
-            key={dependentQuestion.question.name}
-            name={dependentQuestion.question.name}
-            ref={register({
-              ...question.registerConfig,
-              validate: {
-                noEmpty: (value) => value !== '*'
-              }
-            })}
-          >
-            {dependentQuestion.question.config &&
-              dependentQuestion.question.config.options.map((option) => {
-                return (
-                  <option
-                    key={option.value}
-                    value={option.value}
-                    sx={styles.selectOption}
-                  >
-                    {option.label}
-                  </option>
-                )
-              })}
-          </Select>
-        )}
-        {errors[dependentQuestion.question.name] &&
-          (errors[dependentQuestion.question.name].type === 'required' ||
-            errors[dependentQuestion.question.name].type === 'noEmpty') && (
-            <ErrorMessage
-              message={
-                dependentQuestion.question.errorMessages &&
-                dependentQuestion.question.errorMessages.required
-              }
-            />
-          )}
-      </div>
-    )
-  }
-
   return (
     <React.Fragment>
       <div
@@ -91,18 +44,15 @@ const QuestionSelect = ({
           ...(question.isFullWidth && styles.fullWidth)
         }}
       >
-        {question.label && <Label>{question.label}</Label>}
         <Select
-          {...props}
-          defaultValue={getOptions(question)[0]}
-          options={getOptions(question)}
-          isSearchable={false}
           key={question.name}
           name={question.name}
-          register={register}
-          registerConfig={question.registerConfig}
-          setValue={setValue}
-          sx={styles.selectInput}
+          ref={register({
+            ...question.registerConfig,
+            validate: {
+              noEmpty: (value) => value !== '*'
+            }
+          })}
         >
           {question.config &&
             question.config.options.map((option) => {
@@ -127,8 +77,6 @@ const QuestionSelect = ({
             />
           )}
       </div>
-      {question.dependentQuestions &&
-        question.dependentQuestions.map(DependentQuestion)}
     </React.Fragment>
   )
 }
